Created a blocs directory and moved the counter bloc there
