+++ lib/main.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:nesteo_app/bloc/bloc.dart';
+import 'package:nesteo_app/blocs/counter_bloc/counter.dart';
 import 'package:nesteo_app/screens/screens.dart';
 
 void main() => runApp(MyApp());
